tests/requirements.txt
pytest==8.3.4
requests==2.31.0
selenium==4.28.0
webdriver-manager==4.0.2
pytest-html==4.1.1
pytest-cov==5.0.0
pytest-mock==3.14.0
httpx==0.27.2
python-dotenv==1.0.1
